+++ __docs__/wonder-blocks-tabs/navigation-tabs.stories.tsx
@@ -169,7 +169,7 @@ export const HeaderWithNavigationTabsExample: StoryComponentType = {
             navigationTabsRoot: {
                 // set margin to negative value of header vertical spacing so
                 // that selected indicator lines up with header border
-                margin: `-${headerVerticalSpacing} 0`,
+                margin: `calc(${headerVerticalSpacing} * -1) 0`,
             },
         });
         const [currentTab, setCurrentTab] = React.useState(0);