+++ packages/wonder-blocks-testing-core/src/fixtures/fixtures.basic.stories.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-deprecated */
 import * as React from "react";
 
 import {fixtures} from "../index";