+++ packages/wonder-blocks-tokens/src/theme/primitive/sizing.ts
@@ -1,24 +1,4 @@
-/**
- * The baseline for the size tokens.
- */
-const baseline = 10;
-
-/**
- * Converts a number (px) to a rem value.
- */
-function pxToRem(value: number): string {
-    return `${value / baseline}rem`;
-}
-
-/**
- * Converts a rem value to a number (px).
- * @param value The rem value to convert (includes the unit).
- * @returns A string with the px value.
- */
-export function remToPx(value: string): string {
-    const num = parseFloat(value);
-    return `${Math.round(num * baseline)}px`;
-}
+import {pxToRem} from "../../util/sizing-utils";
 
 /**
  * Tokens that define the sizing of elements. These values are expressed in
@@ -44,6 +24,7 @@ export const sizing = {
     size_160: pxToRem(16),
     size_180: pxToRem(18),
     size_200: pxToRem(20),
+    size_220: pxToRem(22),
     size_240: pxToRem(24),
     size_280: pxToRem(28),
     size_320: pxToRem(32),