+++ wb-codemod/src/wb-codemod.js
@@ -30,7 +30,6 @@ async function run(transformFileName, filePaths, options) {
             dry: options.dryRun || false,
             print: options.print || false,
             verbose: 2,
-            babel: true,
             extensions: "js,jsx,ts,tsx",
             parser: "tsx",
             runInBand: !!options.dryRun,