+++ packages/wonder-blocks-accordion/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-accordion
 
+## 3.1.10
+
+### Patch Changes
+
+- Updated dependencies [2656fd4]
+- Updated dependencies [6018552]
+- Updated dependencies [7bbf311]
+- Updated dependencies [7f79943]
+    - @khanacademy/wonder-blocks-tokens@9.0.0
+    - @khanacademy/wonder-blocks-clickable@7.0.3
+
 ## 3.1.9
 
 ### Patch Changes